refactor(ci): supporting git cliff for chosen projects
